// src/others/enums.ts
export enum Role {
    ADMIN = 'ADMIN',
    CLIENT = 'CLIENT'
}
export enum ProductStatus {
    Available = "Available",
    OutOfStock = "Out of Stock",
    Reserved = "Reserved",
    Restocking = "Restocking",
    Discontinued = "Discontinued",
    InTransit = "In Transit",
    Defective = "Defective",
    OnSale = "On Sale",
    UnderReview = "Under Review",
    Withdrawn = "Withdrawn"
}
export enum PaymentMethod {
    CreditCard = "CREDIT_CARD",
    DebitCard = "DEBIT_CARD",
    PayPal = "PAYPAL",
    BankTransfer = "BANK_TRANSFER",
    CashOnDelivery = "CASH_ON_DELIVERY",
    MobilePayment = "MOBILE_PAYMENT",
    Cryptocurrency = "CRYPTOCURRENCY"
}